+++ src/Dictionaries/HashedArrayDictionary.h
@@ -65,14 +65,14 @@ public:
         size_t queries = query_count.load();
         if (!queries)
             return 0;
-        return std::min(1.0, static_cast<double>(found_count.load()) / queries);
+        return std::min(1.0, static_cast<double>(found_count.load()) / static_cast<double>(queries));
     }
 
     double getHitRate() const override { return 1.0; }
 
     size_t getElementCount() const override { return total_element_count; }
 
-    double getLoadFactor() const override { return static_cast<double>(total_element_count) / bucket_count; }
+    double getLoadFactor() const override { return static_cast<double>(total_element_count) / static_cast<double>(bucket_count) ; }
 
     std::shared_ptr<IExternalLoadable> clone() const override
     {
